requirements.txt
Django
django-cors-headers
django-filter
django-simple-captcha
djangorestframework
djangorestframework-simplejwt
requests
pyOpenSSL
psutil
paramiko
channels
daphne
user-agents
django_apscheduler
diskcache
natsort
chardet
redis
qqwry-py3
portalocker
configobj
pywin32 ; sys_platform == 'win32'
PyMySQL
twisted[http2,tls]
setproctitle
pytz
dnspython
cryptography
pycryptodome
pycryptodomex